MR122: Closes 1417265 - Locale problems
I know I promised not to do this again, but it really is all the same thing.

One of the changes between Qt4 and Qt5 is that in Qt4, QString::toDouble()
would convert the string depending on the locale. It does not in Qt5. We have
to use QLocale::toDouble() instead.

This change pretty much does that. Since I was in the metaphorical
neighborhood, I also made sure every call toDouble() checked its return code
and attempts to give a logW as to what couldn't be converted where.

A few notes of interest. Using QObject::property->get() basically calls the
getter defined for that property. Which means you cannot use
QObject::property->get() in the getter. Well, you can. But it loops until you
crash the stack.

+++ src/RecipeExtrasWidget.cpp
@@ -297,7 +297,7 @@ void RecipeExtrasWidget::showChanges(QMetaProperty* prop)
    else if( propName == "tertiaryTemp_c" )
       lineEdit_tertTemp->setText(recipe);
    else if( propName == "tasteRating" )
-      spinBox_tasteRating->setValue( (int)(val.toDouble()) );
+      spinBox_tasteRating->setValue( val.toInt() );
    else if( propName == "date" )
       dateEdit_date->setDate( val.toDate() );
    else if( propName == "notes" )
